Move localStore to utilities
